requirements.txt
Pillow>=11.1.0
pystray>=0.19.5
requests>=2.32.3
pywin32>=305